+++ src/components/Feed/CreateFeed/type.ts
@@ -1,3 +1,12 @@
+import {
+  Control,
+  FieldErrors,
+  UseFormGetValues,
+  UseFormRegister,
+  UseFormSetValue,
+  UseFormWatch,
+} from 'react-hook-form';
+
 export interface FeedType {
   content: string;
   feedImage: string[];
@@ -11,7 +20,20 @@ export interface CreatedFeedTypes {
   modalVisible: boolean;
 }
 
-export interface Inputs {
+export interface FeedFormController {
   content: string;
   feedImage: string[];
+}
+
+export interface FeedTextAreaTypes {
+  errors: FieldErrors<FeedFormController>;
+  register: UseFormRegister<FeedFormController>;
+  watch: UseFormWatch<FeedFormController>;
+}
+
+export interface FeedImageUploadTypes {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  control: Control<FeedFormController, any>;
+  getValues: UseFormGetValues<FeedFormController>;
+  setValue: UseFormSetValue<FeedFormController>;
 }
